+++ lib/src/commands/can.dart
@@ -6,9 +6,8 @@
 
 import 'package:args/command_runner.dart';
 import 'package:gg_log/gg_log.dart';
-import 'package:gg_one/src/commands/can/can_commit.dart';
 import 'package:gg_one/src/commands/can/can_publish.dart';
-import 'package:gg_one/src/commands/can/can_push.dart';
+import 'package:gg_one_commit/gg_one_commit.dart';
 
 // .............................................................................
 /// Various checks for the source code