+++ lib/src/commands/can/can_merge.dart
@@ -7,7 +7,7 @@
 import 'package:gg_one_state/gg_one_state.dart';
 import 'package:gg_args/gg_args.dart';
 import 'package:gg_merge/gg_merge.dart' as gg_merge;
-import 'package:gg_one/src/commands/did/did_commit.dart';
+import 'package:gg_one_commit/gg_one_commit.dart';
 
 /// Are the last changes ready to be merged?
 class CanMerge extends CommandCluster {