+++ lib/src/commands/did.dart
@@ -6,10 +6,9 @@
 
 import 'package:args/command_runner.dart';
 import 'package:gg_log/gg_log.dart';
-import 'package:gg_one/src/commands/did/did_commit.dart';
 import 'package:gg_one/src/commands/did/did_publish.dart';
-import 'package:gg_one/src/commands/did/did_push.dart';
 import 'package:gg_one/src/commands/did/did_upgrade.dart';
+import 'package:gg_one_commit/gg_one_commit.dart';
 
 // .............................................................................
 /// Various checks for the source code